requirements.txt
requests==2.25.1
starkbank-ecdsa==1.1.0
streamlit
newspaper3k
lazyme~=0.0.27
pandas==1.5.2
requests~=2.31.0
nltk~=3.8.1
pycountry==22.3.5
ip2geotools==0.1.6